requirements.txt
requests>=2.31.0
tqdm>=4.66.1
PySimpleGUI>=4.60.5
pyinstaller>=6.1.0
pywin32>=306; sys_platform == 'win32'
pillow>=10.0.0
certifi>=2023.7.22
urllib3>=2.0.7
cryptography>=41.0.4
pyopenssl>=23.2.0
future>=0.18.3
pefile>=2023.2.7
setuptools>=65.5.1
pywin32-ctypes>=0.2.2; sys_platform == 'win32'
pyinstaller-hooks-contrib>=2023.0
virtualenv>=20.0.0
wheel>=0.37.0
fastapi>=0.104.1
uvicorn>=0.24.0
pystray>=0.19.4
python-ffmpeg>=2.0.0
unidecode>=1.3.6
yt-dlp>=2023.11.16
m3u8>=3.6.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
websockets>=11.0.3
python-multipart>=0.0.5
pytest>=7.4.3
aiofiles==23.2.1